Replace content for getauthorization.php
The content and styling for getauthorization.php is replaced for
consistency with the rest of the site. This includes the inclusion
of Google web fonts into getauthorization.php, in accordance to the
rest of the site.
Consider a template header/footer to be included in the future to be
more in line with OOP concepts and increase maintainability.